+++ server/standalone/migration/000039_remove_event_position_default_value.up.sql
@@ -1,0 +1,4 @@
+ALTER TABLE events
+ALTER COLUMN position DROP DEFAULT;
+
+DROP SEQUENCE IF EXISTS seq_events_position;